shelfy_py6/requirements.txt
PySide6
Pillow
requests